// src/wmmt/v388.proto.d.ts
export namespace v388 {

    /** Namespace protobuf. */
    namespace protobuf {
// ...
        namespace SaveGameResultRequest {
// ...
            interface IVersusBattleResult {

                /** VersusBattleResult result */
                result: number;

                /** VersusBattleResult survived */
                survived: boolean;

                /** VersusBattleResult opponentCarId */
                opponentCarId?: (number[]|null);

                /** VersusBattleResult vsPlayCount */
                vsPlayCount: number;

                /** VersusBattleResult vsBurstCount */
                vsBurstCount?: (number|null);

                /** VersusBattleResult vsStarCount */
                vsStarCount?: (number|null);

                /** VersusBattleResult vsCoolOrWild */
                vsCoolOrWild?: (number|null);

                /** VersusBattleResult vsSmoothOrRough */
                vsSmoothOrRough?: (number|null);
            }
// ...
        }
// ...
    }
}